requirements.txt
Firefly-III-API-Client==1.5.4
ImageHash==4.2.1
PyYAML==6.0
numpy==1.22.3
opencv-python==4.5.5.64
pytesseract==0.3.9
python-i18n==0.3.9
python-telegram-bot==13.11
setuptools==62.1.0
price-parser==0.3.4